+++ src/main/java/kitchenpos/application/MenuService.java
@@ -1,20 +1,13 @@
 package kitchenpos.application;
 
-import static kitchenpos.application.validator.MenuValidator.validatePrice;
-import static kitchenpos.application.validator.MenuValidator.validatePriceGreaterThanSum;
 
-import java.math.BigDecimal;
 import java.util.List;
 import java.util.stream.Collectors;
+import kitchenpos.application.validator.MenuValidator;
 import kitchenpos.domain.Menu;
-import kitchenpos.domain.Product;
-import kitchenpos.dto.request.MenuProductRequest;
 import kitchenpos.dto.request.MenuRequest;
 import kitchenpos.dto.response.MenuResponse;
-import kitchenpos.exception.ErrorCode;
-import kitchenpos.exception.KitchenposException;
 import kitchenpos.repository.MenuRepository;
-import kitchenpos.repository.ProductRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -23,22 +16,21 @@ import org.springframework.transaction.annotation.Transactional;
 public class MenuService {
     private final MenuGroupService menuGroupService;
     private final MenuRepository menuRepository;
-    private final ProductRepository productRepository;
+    private final MenuValidator menuValidator;
 
     public MenuService(
             final MenuGroupService menuGroupService,
             final MenuRepository menuRepository,
-            final ProductRepository productRepository
+            final MenuValidator menuValidator
     ) {
         this.menuGroupService = menuGroupService;
         this.menuRepository = menuRepository;
-        this.productRepository = productRepository;
+        this.menuValidator = menuValidator;
     }
 
     public MenuResponse create(final MenuRequest menuRequest) {
         menuGroupService.findById(menuRequest.getMenuGroupId());
-        validatePrice(menuRequest);
-        validatePriceGreaterThanSum(getSumPriceFromMenuProducts(menuRequest.getMenuProducts()), menuRequest.getPrice());
+        menuValidator.validateCreate(menuRequest);
 
         final Menu savedMenu = menuRepository.save(menuRequest.toEntity());
 
@@ -50,17 +42,5 @@ public class MenuService {
         return menuRepository.findAll().stream()
                 .map(MenuResponse::of)
                 .collect(Collectors.toList());
-    }
-
-    private BigDecimal getSumPriceFromMenuProducts(List<MenuProductRequest> menuProducts) {
-        BigDecimal sum = BigDecimal.ZERO;
-
-        for (final MenuProductRequest menuProduct : menuProducts) {
-            final Product product = productRepository.findById(menuProduct.getProductId())
-                    .orElseThrow(() -> new KitchenposException(ErrorCode.NOT_FOUND_PRODUCT));
-            sum = sum.add(product.getPrice().multiply(BigDecimal.valueOf(menuProduct.getQuantity())));
-        }
-
-        return sum;
     }
 }
